chore(contentful): move DummyCMS & ErrorReportingCMS to other files
